CSHARP-2123 Change maxTimeMS set from double to int

+++ src/MongoDB.Driver.Core/Core/Operations/AggregateExplainOperation.cs
@@ -160,7 +160,7 @@ namespace MongoDB.Driver.Core.Operations
                 { "explain", true },
                 { "pipeline", new BsonArray(_pipeline) },
                 { "allowDiskUse", () => _allowDiskUse.Value, _allowDiskUse.HasValue },
-                { "maxTimeMS", () => _maxTime.Value.TotalMilliseconds, _maxTime.HasValue },
+                { "maxTimeMS", () => (int)_maxTime.Value.TotalMilliseconds, _maxTime.HasValue },
                 { "collation", () => _collation.ToBsonDocument(), _collation != null },
                 { "hint", () => _hint, _hint != null },
                 { "comment", () => _comment, _comment != null }
